Fix fullscreen WPT tests assuming fullscreenElement changes each time.
Since fullscreenchange event is sent by the rAF event loop it is possible
that the fullscreenElement might change and not be observed at the time
you read the event. Fix the tests so they are robust.

Fix the ordering of which fullscreen elements are handled in pending
frames. Otherwise we might end up with incorrect ordering.

BUG=874444

Change-Id: I00c228e0aea7794564e522a3883573ea359235f1