Update WebUSB terminate child context tests
This change updates tests for WebUSB where a child context is terminated
while having a USB device open. The existing tests are modified to send
the MojoHandle intercepted by MojoInterfaceInterceptor through a
MessageChannel to allow a new test for a worker context to be created.
As a result, the MojoHandle class is made transferable over a
MessageChannel.

Bug: 841510, 852878
Change-Id: Ib7f74d172b9dcc16a4821a4063cfffc196466a88